Sort on match-depth first for maximum backward compatibility
The current spec favors resource-methods, then sub-resource-methods, and
just then sub-resource locators.  The "enhanced" matching allows to match
any, for example:

    @Path("/") RootResource
        @Path("foo/{bar}") @GET fooAny()
        @Path("foo") : FooResource

    FooResource
        @Path("bar") @GET bar()

    GET /foo/bar : FooResource.bar(),
                   RootResource.fooAny()

By the rule for sorting on number of literal characters first, the
method from the sub-resource class will be prioritized.  To keep it
as backward-compatible as possible, sort on match depth first (smallest
first).  Of course users should reasonably disambiguate their @Path
declarations, like:

    @Path("/") RootResource
        @Path("foo/{bar:(?!bar/?)}") @GET fooAny()
        @Path("foo") : FooResource

    FooResource
        @Path("bar") @GET bar()

I'm still not sure this is the right thing to do, as I see greater
utility in the previous "full path match" sorting behavior.  Using any
of the above examples is not possible with the current JAX-RS, anyway.